requirements.txt
azure-functions
pymysql
numpy
scipy
nltk
gensim
logging